DAS-NONE: Move trajectory subsetter tests to shared_utils (#104)
