+++ README.md
@@ -222,7 +222,7 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - python=3.7
+  - python=3.11
   - jupyter
   - requests
   - netcdf4
