+++ test/trajectory-subsetter/version.txt
@@ -1,1 +1,1 @@
-0.1.6
+0.1.7
